+++ src/tools/contract.cpp
@@ -12,6 +12,7 @@
 #include <cstdlib>
 #include <filesystem>
 #include <new>
+#include <set>
 #include <thread>
 
 #include "util/meminfo.hpp"
@@ -33,6 +34,7 @@ return_code parseArguments(int argc,
     // declare a group of options that will be allowed only on command line
     boost::program_options::options_description generic_options("Options");
     generic_options.add_options()("version,v", "Show version")("help,h", "Show this help message")(
+        "list-inputs", "List required and optional input file extensions")(
         "verbosity,l",
         boost::program_options::value<std::string>(&verbosity)->default_value("INFO"),
         std::string("Log verbosity level: " + util::LogPolicy::GetLevels()).c_str());
@@ -117,6 +119,15 @@ return_code parseArguments(int argc,
     if (option_variables.count("help"))
     {
         std::cout << visible_options;
+        return return_code::exit;
+    }
+
+    if (option_variables.count("list-inputs"))
+    {
+        contractor::ContractorConfig config;
+        std::set<std::string> seen;
+        config.ListInputFiles(std::cout, seen);
+        config.updater_config.ListInputFiles(std::cout, seen);
         return return_code::exit;
     }